Move VoteTransaction to vote crate
