+++ core/src/consensus.rs
@@ -34,13 +34,13 @@ use {
         signature::Keypair,
         slot_history::{Check, SlotHistory},
     },
-    solana_vote::vote_account::VoteAccountsHashMap,
+    solana_vote::{vote_account::VoteAccountsHashMap, vote_transaction::VoteTransaction},
     solana_vote_program::{
         vote_error::VoteError,
         vote_instruction,
         vote_state::{
             process_slot_vote_unchecked, BlockTimestamp, Lockout, TowerSync, Vote,
-            VoteState1_14_11, VoteStateUpdate, VoteTransaction, MAX_LOCKOUT_HISTORY,
+            VoteState1_14_11, VoteStateUpdate, MAX_LOCKOUT_HISTORY,
         },
     },
     std::{
