+++ core/src/consensus/tower_storage.rs
@@ -381,9 +381,9 @@ pub mod test {
             BlockhashStatus, Tower,
         },
         solana_sdk::{hash::Hash, signature::Keypair},
+        solana_vote::vote_transaction::VoteTransaction,
         solana_vote_program::vote_state::{
-            BlockTimestamp, LandedVote, Vote, VoteState, VoteState1_14_11, VoteTransaction,
-            MAX_LOCKOUT_HISTORY,
+            BlockTimestamp, LandedVote, Vote, VoteState, VoteState1_14_11, MAX_LOCKOUT_HISTORY,
         },
         tempfile::TempDir,
     };
