+++ programs/vote/src/lib.rs
@@ -2,13 +2,11 @@
 
 pub mod vote_processor;
 pub mod vote_state;
-pub mod vote_transaction;
 
 #[cfg_attr(feature = "metrics", macro_use)]
 #[cfg(feature = "metrics")]
 extern crate solana_metrics;
 
-#[cfg_attr(feature = "frozen-abi", macro_use)]
 #[cfg(feature = "frozen-abi")]
 extern crate solana_frozen_abi_macro;
