+++ programs/vote/src/vote_state/mod.rs
@@ -3,9 +3,8 @@
 pub use solana_vote_interface::state::{vote_state_versions::*, *};
 use {
     log::*,
-    serde_derive::{Deserialize, Serialize},
     solana_account::{AccountSharedData, ReadableAccount, WritableAccount},
-    solana_clock::{Clock, Epoch, Slot, UnixTimestamp},
+    solana_clock::{Clock, Epoch, Slot},
     solana_epoch_schedule::EpochSchedule,
     solana_feature_set::{self as feature_set, FeatureSet},
     solana_hash::Hash,
@@ -20,132 +19,8 @@ use {
     std::{
         cmp::Ordering,
         collections::{HashSet, VecDeque},
-        fmt::Debug,
     },
 };
-
-#[cfg_attr(
-    feature = "frozen-abi",
-    derive(AbiExample, AbiEnumVisitor),
-    frozen_abi(digest = "4BdRo6We16yDbjz69H1oFq6C4nXUZKDchZR76TvvGmBi")
-)]
-#[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
-pub enum VoteTransaction {
-    Vote(Vote),
-    VoteStateUpdate(VoteStateUpdate),
-    #[serde(with = "serde_compact_vote_state_update")]
-    CompactVoteStateUpdate(VoteStateUpdate),
-    #[serde(with = "serde_tower_sync")]
-    TowerSync(TowerSync),
-}
-
-impl VoteTransaction {
-    pub fn slots(&self) -> Vec<Slot> {
-        match self {
-            VoteTransaction::Vote(vote) => vote.slots.clone(),
-            VoteTransaction::VoteStateUpdate(vote_state_update) => vote_state_update.slots(),
-            VoteTransaction::CompactVoteStateUpdate(vote_state_update) => vote_state_update.slots(),
-            VoteTransaction::TowerSync(tower_sync) => tower_sync.slots(),
-        }
-    }
-
-    pub fn slot(&self, i: usize) -> Slot {
-        match self {
-            VoteTransaction::Vote(vote) => vote.slots[i],
-            VoteTransaction::VoteStateUpdate(vote_state_update)
-            | VoteTransaction::CompactVoteStateUpdate(vote_state_update) => {
-                vote_state_update.lockouts[i].slot()
-            }
-            VoteTransaction::TowerSync(tower_sync) => tower_sync.lockouts[i].slot(),
-        }
-    }
-
-    pub fn len(&self) -> usize {
-        match self {
-            VoteTransaction::Vote(vote) => vote.slots.len(),
-            VoteTransaction::VoteStateUpdate(vote_state_update)
-            | VoteTransaction::CompactVoteStateUpdate(vote_state_update) => {
-                vote_state_update.lockouts.len()
-            }
-            VoteTransaction::TowerSync(tower_sync) => tower_sync.lockouts.len(),
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        match self {
-            VoteTransaction::Vote(vote) => vote.slots.is_empty(),
-            VoteTransaction::VoteStateUpdate(vote_state_update)
-            | VoteTransaction::CompactVoteStateUpdate(vote_state_update) => {
-                vote_state_update.lockouts.is_empty()
-            }
-            VoteTransaction::TowerSync(tower_sync) => tower_sync.lockouts.is_empty(),
-        }
-    }
-
-    pub fn hash(&self) -> Hash {
-        match self {
-            VoteTransaction::Vote(vote) => vote.hash,
-            VoteTransaction::VoteStateUpdate(vote_state_update) => vote_state_update.hash,
-            VoteTransaction::CompactVoteStateUpdate(vote_state_update) => vote_state_update.hash,
-            VoteTransaction::TowerSync(tower_sync) => tower_sync.hash,
-        }
-    }
-
-    pub fn timestamp(&self) -> Option<UnixTimestamp> {
-        match self {
-            VoteTransaction::Vote(vote) => vote.timestamp,
-            VoteTransaction::VoteStateUpdate(vote_state_update)
-            | VoteTransaction::CompactVoteStateUpdate(vote_state_update) => {
-                vote_state_update.timestamp
-            }
-            VoteTransaction::TowerSync(tower_sync) => tower_sync.timestamp,
-        }
-    }
-
-    pub fn set_timestamp(&mut self, ts: Option<UnixTimestamp>) {
-        match self {
-            VoteTransaction::Vote(vote) => vote.timestamp = ts,
-            VoteTransaction::VoteStateUpdate(vote_state_update)
-            | VoteTransaction::CompactVoteStateUpdate(vote_state_update) => {
-                vote_state_update.timestamp = ts
-            }
-            VoteTransaction::TowerSync(tower_sync) => tower_sync.timestamp = ts,
-        }
-    }
-
-    pub fn last_voted_slot(&self) -> Option<Slot> {
-        match self {
-            VoteTransaction::Vote(vote) => vote.last_voted_slot(),
-            VoteTransaction::VoteStateUpdate(vote_state_update)
-            | VoteTransaction::CompactVoteStateUpdate(vote_state_update) => {
-                vote_state_update.last_voted_slot()
-            }
-            VoteTransaction::TowerSync(tower_sync) => tower_sync.last_voted_slot(),
-        }
-    }
-
-    pub fn last_voted_slot_hash(&self) -> Option<(Slot, Hash)> {
-        Some((self.last_voted_slot()?, self.hash()))
-    }
-}
-
-impl From<Vote> for VoteTransaction {
-    fn from(vote: Vote) -> Self {
-        VoteTransaction::Vote(vote)
-    }
-}
-
-impl From<VoteStateUpdate> for VoteTransaction {
-    fn from(vote_state_update: VoteStateUpdate) -> Self {
-        VoteTransaction::VoteStateUpdate(vote_state_update)
-    }
-}
-
-impl From<TowerSync> for VoteTransaction {
-    fn from(tower_sync: TowerSync) -> Self {
-        VoteTransaction::TowerSync(tower_sync)
-    }
-}
 
 // utility function, used by Stakes, tests
 pub fn from<T: ReadableAccount>(account: &T) -> Option<VoteState> {
