+++ vote/src/lib.rs
@@ -5,9 +5,9 @@ pub mod vote_account;
 pub mod vote_parser;
 pub mod vote_transaction;
 
-#[macro_use]
-extern crate serde_derive;
-
 #[cfg_attr(feature = "frozen-abi", macro_use)]
 #[cfg(feature = "frozen-abi")]
 extern crate solana_frozen_abi_macro;
+
+#[macro_use]
+extern crate serde_derive;
